requirements.txt
datasets
pandas
tqdm
langchain_huggingface
langchain_text_splitters
langchain_community
langchain_core
huggingface_hub
python-dotenv
pyarrow
faiss-cpu
langchain_google_genai
